Feature: Core Civic Enhancements

+++ Nagar-Seva/backend/src/main/java/com/nagarseva/service/AuditService.java
@@ -152,7 +152,7 @@ public class AuditService {
                                     String department, Object oldConfig, Object newConfig,
                                     HttpServletRequest request) {
         String compositeKey = category + "|" + department;
-        Long entityId = compositeKey.hashCode();
+        Long entityId = (long) compositeKey.hashCode();
         log(actorEmail, actorRole, "SLA_CONFIG_CHANGED", "SlaConfig", entityId,
                 oldConfig, newConfig, request, true, null);
     }
